test: 급여 설정 수정 (#20)
